% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newXOR.R
\docType{data}
\name{envXOR}
\alias{envXOR}
\title{The problem environment \code{envXOR} for programming the XOR function
either by grammar-based genetic programming or grammatical evolution.}
\format{
An object of class \code{list} of length 4.
}
\usage{
envXOR
}
\description{
The problem environment \code{envXOR} is a list with the following elements:
\itemize{
\item \code{envXOR$name}: 
        \code{"envXOR"}, the name of the problem environment.
\item \code{envXOR$buildtest(expr)}: 
                              The function which builds the environment 
                              for evaluating the expression with 
                              a binding of the variables to the parameters.
\item \code{envXOR$TestCases}: The truth table of the XOR function.
\item \code{envXOR$f(expr, gene=NULL, lF=NULL)}: The fitness function.
                             \code{expr} is the string with the 
                             logical expression to be evaluated. 
}
}
\seealso{
Other Problem Environments: 
\code{\link{DeJongF4Factory}()},
\code{\link{DelayedPFactory}()},
\code{\link{Parabola2DEarlyFactory}()},
\code{\link{Parabola2DErrFactory}()},
\code{\link{Parabola2DFactory}()},
\code{\link{lau15}},
\code{\link{newEnvXOR}()},
\code{\link{newTSP}()}
}
\concept{Problem Environments}
\keyword{datasets}
