% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalGene.R
\name{NewlFevalGenes}
\alias{NewlFevalGenes}
\title{Generate local functions and objects}
\usage{
NewlFevalGenes(penv)
}
\arguments{
\item{penv}{A problem environment.}
}
\value{
The local configuration. A list of functions.
}
\description{
\code{NewlFevalGenes()} returns 
             the list of functions containing
             a definition of all local objects required for the use
             of evaluation functions. We reference this object 
             as local configuration. When adding additional 
             evaluation functions, this must be extended
             by the constant (functions) needed to configure them.
}
\examples{
Parabola2D<-Parabola2DFactory()
lF<-NewlFevalGenes(Parabola2D)
lF$Max()
}
