% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGene.R
\name{NewlFselectGenes}
\alias{NewlFselectGenes}
\title{Generate local functions and objects.}
\usage{
NewlFselectGenes()
}
\value{
Local configuration \code{lF}.
}
\description{
\code{NewlFselectGenes()} returns 
             the list of functions which contains 
             a definition of all local objects required for the use
             of selection functions. We reference this object 
             as local configuration. When adding additional 
             selection functions, this must be extended
             by the constant (functions) needed to configure them.
}
\examples{
lF<-NewlFselectGenes()
lF$Max()
}
