% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalGene.R
\name{testEvalGeneStoch}
\alias{testEvalGeneStoch}
\title{Test of incremental mean, variance, and standard deviation.}
\usage{
testEvalGeneStoch(gene, lF, rep)
}
\arguments{
\item{gene}{A gene.}

\item{lF}{A local function list with a problem environment.}

\item{rep}{Number of repeated evaluations.}
}
\value{
A gene.
}
\description{
Test of incremental mean, variance, and standard deviation.
}
\examples{
DeJongF4<-DeJongF4Factory()
lF<-NewlFevalGenes(DeJongF4)
g1<-list(evaluated=FALSE, evalFail=FALSE, fit=0, gene1=c(1.0, -1.5))
g10<-testEvalGeneStoch(g1, lF, rep=10)
g10
}
\concept{Tests}
