% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGene.R
\name{STournament}
\alias{STournament}
\title{Stochastic tournament of size \code{k}.}
\usage{
STournament(fit, lF)
}
\arguments{
\item{fit}{Fitness vector.}

\item{lF}{Local configuration.}
}
\value{
Index of candidate.
}
\description{
\code{STournament()} is implemented in two steps:
\enumerate{
\item
A subset of size k of the population is selected with uniform probability.
\item
A gene is selected with probability proportional to fitness.
}
}
\examples{
fit<-sample(10, 15, replace=TRUE)
STournament(fit, NewlFselectGenes()) 
}
